+++ lib/circulito.dart
@@ -5,9 +5,11 @@ import 'package:flutter/material.dart';
 import 'src/circulito_painter.dart';
 import 'src/circulito_section.dart';
 import 'src/circulito_stroke_cap.dart';
+import 'src/start_point.dart';
 
 export 'src/circulito_section.dart';
 export 'src/circulito_stroke_cap.dart';
+export 'src/start_point.dart';
 
 /// Circulito is a widget wraps the CirculitoPainter class
 /// to be used properly.
@@ -52,6 +54,11 @@ class Circulito extends StatelessWidget {
   /// Could be `Butt` or `Round`. Default to `Round`.
   final CirculitoStrokeCap strokeCap;
 
+  /// Determines the start point of the wheel.
+  ///
+  /// Could be `top`, `bottom`, `left` or `right`. Default to `top`.
+  final StartPoint startPoint;
+
   const Circulito({
     super.key,
     required this.sections,
@@ -60,6 +67,7 @@ class Circulito extends StatelessWidget {
     this.backgroundColor,
     this.strokeWidth = 20,
     this.isCentered = true,
+    this.startPoint = StartPoint.top,
     this.strokeCap = CirculitoStrokeCap.round,
   })  : assert(strokeWidth > 0, "[strokeWidth] must be a positive value"),
         assert(maxSize > 0, "[maxSize] must be a positive value"),
@@ -77,6 +85,7 @@ class Circulito extends StatelessWidget {
         strokeWidth: strokeWidth,
         isCentered: isCentered,
         strokeCap: strokeCap,
+        startPoint: startPoint,
         backgroundColor: backgroundColor,
       ),
     );